+++ src/Confluent.Kafka/GroupInfo.cs
