+++ src/Confluent.Kafka/Offset.cs
