+++ src/Confluent.Kafka/Timestamp.cs
