+++ src/Confluent.Kafka/TopicMetadata.cs
@@ -16,25 +16,37 @@
 //
 // Refer to LICENSE for more information.
 
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+
 namespace Confluent.Kafka
 {
+
     /// <summary>
-    ///     Metadata pertaining to a single Kafka broker.
+    ///     Metadata pertaining to a single Kafka topic.
     /// </summary>
-    public struct BrokerMetadata
+    public class TopicMetadata
     {
-        public BrokerMetadata(int brokerId, string host, int port)
+        public TopicMetadata(string topic, List<PartitionMetadata> partitions, Error error)
         {
-            BrokerId = brokerId;
-            Host = host;
-            Port = port;
+            Topic = topic;
+            Error = error;
+            Partitions = partitions;
         }
 
-        public int BrokerId { get; }
-        public string Host { get; }
-        public int Port { get; }
+        public string Topic { get; }
+        public List<PartitionMetadata> Partitions { get; }
+        public Error Error { get; }
 
         public override string ToString()
-            => $"{{ \"BrokerId\": {BrokerId}, \"Host\": \"{Host}\", \"Port\": {Port} }}";
+        {
+            var result = new StringBuilder();
+            result.Append($"{{ \"Topic\": \"{Topic}\", \"Partitions\": [");
+            result.Append(string.Join(",", Partitions.Select(p => $" {p.ToString()}")));
+            result.Append($" ], \"Error\": \"{Error.Code.ToString()}\" }}");
+            return result.ToString();
+        }
     }
 }
