+++ test/Confluent.Kafka.IntegrationTests/Tests/ProduceTimestamp.cs
@@ -49,6 +49,7 @@ namespace Confluent.Kafka.IntegrationTests
             using (var producer = new Producer<Null, string>(producerConfig, null, new StringSerializer(Encoding.UTF8)))
             {
                 dr = producer.ProduceAsync(topic, null, testString, testTime).Result;
+                Assert.NotNull(dr);
                 Assert.Equal(topic, dr.Topic);
                 Assert.NotEqual<long>(dr.Offset, Offset.Invalid);
                 Assert.Equal(dr.Timestamp.DateTime, testTime);
@@ -61,6 +62,7 @@ namespace Confluent.Kafka.IntegrationTests
                 consumer.Assign(new List<TopicPartitionOffset>() { dr.TopicPartitionOffset });
                 Message msg;
                 Assert.True(consumer.Consume(out msg, TimeSpan.FromSeconds(10)));
+                Assert.NotNull(msg);
                 Assert.Equal(msg.Timestamp.DateTime, testTime);
                 Assert.True(false); // TODO: check timestamp type. what should it be?
             }
