+++ test/Confluent.Kafka.IntegrationTests/Util.cs
@@ -43,12 +43,13 @@ namespace Confluent.Kafka.IntegrationTests
             }
             var msg = sb.ToString();
 
-            Message<Null, string> firstDeliveryReport = default(Message<Null, string>);
+            Message<Null, string> firstDeliveryReport = null;
             using (var producer = new Producer<Null, string>(producerConfig, null, new StringSerializer(Encoding.UTF8)))
             {
                 for (int i=0; i<number; ++i)
                 {
                     var dr = producer.ProduceAsync(topic, null, msg).Result;
+                    Assert.NotNull(dr);
                     if (i == 0)
                     {
                         firstDeliveryReport = dr;
